### 2023-10-25 #5985 @editorialbot
[PRE REVIEW]: PyTASER: Simulating transient absorption spectroscopy (TAS) for crystals from first principles
**Submitting author:** @savya10 (<a href="http://orcid.org/0009-0007-7128-3465">Savyasanchi Aggarwal</a>)
**Repository:** https://github.com/WMD-group/PyTASER
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.2.0
