### 2023-10-25 #5985 @Kevin-Mattheus-Moerman
@savya10 thanks for this submission. I am the AEiC on this track and here to help with initial steps.

It looks like your paper is not compiling because rather than just a newline on line 37, you have a "tab" "character". If you replace that line an actual empty line I think that might fix it. Can you update the paper? Once you have done that, you can call `@editorialbot generate pdf` here to retry the paper compilation.
