+++ paper/paper.md
@@ -34,7 +34,7 @@ authors:
     orcid: 0000-0001-5460-7033
     corresponding: true
     affiliation: "1"
-	
+
 affiliations:
  - name: Thomas Young Centre and Department of Materials, Imperial College London, London, United Kingdom
    index: 1
