### 2023-10-30 #5999 @editorialbot
[REVIEW]: PyTASER: Simulating transient absorption spectroscopy (TAS) for crystals from first principles
**Submitting author:** @savya10 (<a href="http://orcid.org/0009-0007-7128-3465">Savyasanchi Aggarwal</a>)
**Repository:** https://github.com/WMD-group/PyTASER
**Branch with paper.md** (empty if default branch): paper
**Version:** 2.3.0
