### 2023-11-20 #5999 @obaica
Hi @jgostick , I've finished my review. PyTASER is a well-designed and versatile tool that represents a significant advancement in the field of materials science, particularly in the study of transient (TAS) and differential (DAS) absorption spectroscopies. This software is capable of not only processing outputs from density functional theory (DFT) calculations but also utilizing data from the (legacy) Materials Project database. It can generate predictive TAS under various conditions, such as changes in temperature and carrier concentration. I highly recommend its publication in JOSS, as it will undoubtedly be a valuable asset to the scientific community.

For the authors @savya10 et al., some comments are listed below.

- **Issue with `conc` Parameter in TAS Generation**: When altering the `conc` parameter in `tas = tg.generate_tas(energy_min=0, energy_max=7, temp=300, conc=1e21, cshift=1e-3)` to values below 1e15, I noticed that the title in the plots generated by `from pytaser.plotter import TASPlotter` does not display `conc` in scientific notation. I recommend modifying the following section to format the output of `conc` more effectively:
```
        plt.title(
                    abs_label
                    + " spectrum of "
                    + formatted_material_name
                    + " at T = "
                    + str(self.temp)
                    + " K, n = "
                    + str(self.conc)
                    + " $cm^{-3}$",
                    fontsize=25,
                )
```
- **Compatibility with New pymatgen Versions**: In the `PyTASER Materials Project Workflow` example, it seems that with newer versions of `pymatgen`, `from pymatgen.ext.matproj import TaskType` should be replaced with `from pymatgen.ext.matproj_legacy import TaskType` for proper functionality.
- **Enhanced Documentation on Band Contributions**:  I suggest the manual should include more comprehensive descriptions regarding the individual band-to-band transitions observed in the examples. Specifically, elucidating the meaning of `labels` for band pairs in the spectrum, such as the notation “(-1,2)、(0,1)”, would be highly beneficial for a non-specialist audience.
- **Additional Examples for Customizable Plotting**: Incorporating more examples on customizable plotting in the manual, such as settings for image size and output format, would undoubtedly enhance the practicality of the software.
- **Inclusion of Experimental Data Comparison**: Given that the goal of this library is to simulate spectra for comparison with and interpretation of experiments, adding examples and explanations that compare with experimental data would significantly increase the software's power, especially for a non-specialist audience.

Thank you for your patience as I took the time to thoroughly prepare this review.
