### 2023-12-12 #5999 @JosePizarro3
Hi @savya10 @jgostick

I chained conferences with vacations, everything before Christmas. You can see the gap in my Github contributions timeline 😄

Sadly, I didn't have time to play with the package. I also predicted this in my initial emails, and communicated that by (~mid) January you will have my review. Sorry I cannot give you a more satisfactory answer as of now, but I hope you understand 🙂
