### 2024-01-08 #5999 @JosePizarro3
Hi @jgostick @savya10,

Thank you very much for the patience.

I finished my review, and I have to say that this software is very interesting. I pretty much agree with @obaica's comments. I would like to stress a bit more how cool is that PyTASER can recognize first principles outputs (so far VASP and MP, but who knows, maybe in the future some more electronic structure codes and other databases?) and generate TAS and DAS 🙂 I am a big fan of scientific softwares that are interoperable with different existing codes. Furthermore, it is very nice we can indeed pass different parameters as inputs, and even fix the DFT reported band gap to calculate these spectras.

I have some comments, feedback, and questions for the authors:

1. I think adding `setuptools` in the requirements.txt is an unsafe practice, as it might enter in conflict with other packages, most importantly, with `pip`. Maybe you want to check the installation.
2. Furthermore, I was not able to install the package using `setup.py`. I had some troubles with `spglib` if pip is not upgraded prior to the installation, but then later with `ruamel.yaml`. This last error I was not able to fix it easily. Maybe it is a problem of the machine I used for installing the package, but I am also puzzled on why installing with `pip` worked, while the `setup.py` didn't.
3. I think you can improve a bit on the documentation (specially on the introduction and where to give installation instructions), while also leaving out some larger changes for a future version. I felt some information was repeated in the welcoming and under the Information menu.
4. I think you could directly add the nice statement of need from the paper into the documentation.
5. Can you please add "dark" and "light" labels in Figure 1 of the paper? I think you could also add this figure also in the documentation page, as an introduction of the "dark" and "light" concepts and what an user can expect when outputting spectras.
6. `pytaser/__init__.py` contains a different `__version__` than the package one. I would simply delete this, as it forces you to keep track of the version tag there, but like you prefer.
7. There are a couple of imports not being used in `pytaser/plotter.py` (`gettext` and `collections`). I am surprised neither `black` nor `flake8` caught this. Did you consider to maybe use some linting checks on the github pipeline? If so, I would suggest [`ruff`](https://docs.astral.sh/ruff/).
8. I like the testing, but on first sight it seems you are testing `generator.py`, `tas.py` and `plotter.py`. Maybe you can add some other tests for the other modules in a future version. I would also suggest you add some docstrings to the testing, so it is clear which methods is covering each one of them.
9. I think the order of the x axis in the plots is a bit confusing. This is because sometimes it is inverted and sometimes not, even thought the x axis might be using Energies or Wavelengths (see e.g., cells [5] and [10] on the DFT example). I would suggest to add the `gca().invert_xaxis()` as an input flag of `get_plot()`. If this is not specified, then the method should be able to handle the fact that Energies go from smaller to larger values, while Wavelengths from larger to smaller values.
10. Finally, I think you should leave more interactivity on the plotting of the spectra. I would say that cell [23] of the DFT example is a good case for this. Maybe it is a good idea to leverage how many contributions and how many lines you want to show for a given spectra. The colors and the lines should be also checked on this (see again the cell [23]).

Besides these, I think the code is very clean and nicely written, congratulations! 😄
