### 2024-01-08 #5999 @savya10
Hi @JosePizarro3, thanks very much for the comprehensive review and nice feedback! We'll implement the helpful changes that both you and @obaica mentioned and will release an updated version soon.
Regarding leaving some more of the background/theory in the documentation, we will shortly be releasing a separate paper going into more detail about this, but will certainly add some more into the current docs if you believe it will help.

Thanks again, and happy new year to all! 😁
