### 2024-01-22 #5999 @jgostick
Hi @savya10, the editorialbot has suggested some DOIs for the citations that are missing some.  Though it missed "Perovskite-147 inspired materials for photovoltaics and beyond—from design to devices". Please update the paper.md file.
