### 2024-01-22 #5999 @jgostick
My checklist also suggests that I look for inconsistencies in the bibliography.  Most things seem ok, but here are few suggestions:
- The reference to "On the quantization of the monoatomic ideal gas" is actually a translation of a Fermi paper correct?  Would it be more correct to actually cite the original?
- The reference to "Dresselhaus, M. S. (2001)" is to lecture notes...is there a more generally available source for this material, or are these lecture notes considered "part of the literature" in the field?
- The capitalization of the article titles is not consistent. Most are in normal sentence case, but a few have each word capitalized.
