+++ paper/paper.md
@@ -68,7 +68,7 @@ The drawback of modern TAS is that the spectra are often difficult to interpret 
 
 `PyTASER` identifies the allowed vertical optical transitions between the electronic bands of the material to determine the possible excitations that can occur in the respective ground 'dark' and excited 'light' stages. It does this by calculating the effective absorption for each state; a product of the joint density of states (JDOS) in the material and the transition probability for each band transition. These are based on post-processing of ground-state density functional theory calculations. Once calculated, `PyTASER` then compares the change in electronic transitions between the dark and light states. 
 
-![Schematics of the ground and excited state electronic structures and optical profiles. The ground 'dark' state is at the top, showing full occupancy and unoccupancy (blue, orange) for the conduction and valence bands respectively. The excited `light` state shows partial occupancy in a similar plot at the bottom. The overall DA plot is displayed to the right, the difference between the dark and light effective absorption plots. \label{fig:figure1}](Fig1.pdf){width=100mm}
+![Schematics of the ground and excited state electronic structures and optical profiles. The ground 'dark' state is at the top, showing full occupancy and unoccupancy (blue, orange) for the conduction and valence bands respectively. The excited 'light' state shows partial occupancy in a similar plot at the bottom. The overall DA plot is displayed to the right, the difference between the dark and light effective absorption plots. \label{fig:figure1}](Fig1.pdf){width=100mm}
 
 ## JDOS method
 
@@ -80,7 +80,7 @@ JDOS is defined as the density of allowed vertical band-to-band transitions base
 Here, $c$ and $v$ refer to the conduction and valence bands respectively. $\varepsilon$ refers to the energy of the respective band at kpoint $\boldsymbol{k}$.
 
 Determining the JDOS for the light state is more difficult, as the initial 'pump' excitation leads to partial occupancies in both the valence and conduction bands, which can contribute to additional optical transitions within these bands.
-`PyTASER` uses quasi-Fermi levels [@nelson2003physics; @dresselhaus2001solid] to address such band transitions, deviating from specific valence-to-conduction band transitions to favour initial-to-final band transitions (\autoref{eq:jdos_pytaser}). The partial occupancies ($f_{i,k}$ and $f_{f,k}$ in \autoref{eq:jdos_pytaser}) centred at these quasi-Fermi levels can be estimated by using the Fermi-Dirac distribution [@zannoni1999quantization; @dirac1926theory] as the light excitation introduces excess charge carriers (holes and electrons) into the material.
+`PyTASER` uses quasi-Fermi levels [@nelson2003physics; @dresselhaus2001solid] to address such band transitions, deviating from specific valence-to-conduction band transitions to favour initial-to-final band transitions (\autoref{eq:jdos_pytaser}). The partial occupancies ($f_{i,k}$ and $f_{f,k}$ in \autoref{eq:jdos_pytaser}) centred at these quasi-Fermi levels can be estimated by using the Fermi-Dirac distribution [@fermi1926sulla; @dirac1926theory] as the light excitation introduces excess charge carriers (holes and electrons) into the material.
 The use of Fermi-Dirac statistics introduces two variables; the effective temperature and concentration of free carriers in the material. The latter is related to the strength of the initial pump, as well as the pump-probe time delay. These can be used to understand the time-evolution of the excited state in the material.
 
 \begin{equation}
