+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{clarke:2010,
-	title = {Charge {Photogeneration} in {Organic} {Solar} {Cells}},
+	title = {Charge photogeneration in organic solar cells},
 	volume = {110},
 	issn = {0009-2665},
 	url = {https://doi.org/10.1021/cr900271s},
@@ -151,7 +151,7 @@ Program Title: Abinit Program Files doi: http://dx.doi.org/10.17632/csvdrr4d68.1
 }
 
 @article{wang:2015,
-	title = {Transient {Absorption} {Spectroscopy} of {Anatase} and {Rutile}: {The} {Impact} of {Morphology} and {Phase} on {Photocatalytic} {Activity}},
+	title = {Transient Absorption Spectroscopy of Anatase and Rutile: The impact of morphology and phase on photocatalytic activity},
 	volume = {119},
 	issn = {1932-7447},
 	shorttitle = {Transient {Absorption} {Spectroscopy} of {Anatase} and {Rutile}},
@@ -170,7 +170,7 @@ Program Title: Abinit Program Files doi: http://dx.doi.org/10.17632/csvdrr4d68.1
 }
 
 @article{kafizas:2016,
-     title = {Where {Do} {Photogenerated} {Holes} {Go} in {Anatase}:{Rutile} {TiO2}? {A} {Transient} {Absorption} {Spectroscopy}        {Study} of {Charge} {Transfer} and {Lifetime}},
+     title = { Where do photogenerated holes go in Anatase:Rutile TiO2? A Transient Absorption Spectroscopy study of charge transfer and lifetime},
      volume = {120},
      issn = {1089-5639},
      shorttitle = {Where {Do} {Photogenerated} {Holes} {Go} in {Anatase}},
@@ -192,16 +192,16 @@ Program Title: Abinit Program Files doi: http://dx.doi.org/10.17632/csvdrr4d68.1
   title={The physics of solar cells},
   author={Nelson, Jenny A},
   year={2003},
-  publisher={World Scientific Publishing Company}
+  publisher={World Scientific Publishing Company},
+  doi={10.1142/p276}
 }
 
-@misc{zannoni1999quantization,
-      title={On the Quantization of the Monoatomic Ideal Gas}, 
-      author={Alberto Zannoni},
-      year={1999},
-      eprint={cond-mat/9912229},
-      archivePrefix={arXiv},
-      primaryClass={cond-mat.stat-mech}
+@article{fermi1926sulla,
+  title={Sulla quantizzazione del gas perfetto monoatomico},
+  author={Fermi, Enrico},
+  journal={Rendiconti Lincei},
+  volume={145},
+  year={1926}
 }
 
 @article{dirac1926theory,
@@ -212,7 +212,8 @@ Program Title: Abinit Program Files doi: http://dx.doi.org/10.17632/csvdrr4d68.1
   number={762},
   pages={661--677},
   year={1926},
-  publisher={The Royal Society London}
+  publisher={The Royal Society London},
+  doi={10.1098/rspa.1926.0133}
 }
 
  @article{dresselhaus2001solid,
@@ -227,6 +228,7 @@ Program Title: Abinit Program Files doi: http://dx.doi.org/10.17632/csvdrr4d68.1
 @article{huang2021perovskite,
   title={Perovskite-inspired materials for photovoltaics and beyond—from design to devices},
   author={Huang, Yi-Teng and Kavanagh, Se{\'a}n R and Scanlon, David O and Walsh, Aron and Hoye, Robert LZ},
+  doi={10.1088/1361-6528/abcf6d}
   journal={Nanotechnology},
   volume={32},
   number={13},
[…]
