### 2024-01-31 df3c45e
Update `sleep` command in `pip install test` to work on Windows and Mac
