+++ .github/workflows/pip_install_test.yml
@@ -29,7 +29,7 @@ jobs:
 
       - name: Install dependencies
         run: |
-          sleep 10m  # wait 10 minutes for PyPI to update with the new release
+          sleep 600  # wait 10 minutes for PyPI to update with the new release
           python -m pip install --upgrade pip
           pip install pytaser  # install only from PyPI
           pip install pytaser[tests]
