### 2024-01-31 ece8236
Update `python setup.py install` instructions
