### 2024-01-31 925b6e2
Remove unnecessary `setuptools` requirement (JosePizarro3 review Q1)
