+++ requirements.txt
@@ -5,5 +5,4 @@ monty
 pymatgen>=2023.05.31
 matplotlib>=3.7.1
 scipy
-setuptools
 pathlib
