+++ pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
 # Minimum requirements for the build system to execute.
-requires = ["setuptools"]  # PEP 508 specifications.
+requires = ["setuptools>=42"]  # PEP 508 specifications.
 build-backend = "setuptools.build_meta"
 
 [tool.semantic_release]
