### 2024-01-31 acb666a
Remove `__version__` from `__init__.py` (JosePizarro3 review Q6)
