+++ pytaser/__init__.py
@@ -1,3 +1,1 @@
 """PyTASER is a package for the generation, analysis and plotting of Transient Absorption Spectra of bulk materials."""
-
-__version__ = "2.1.0"
