### 2024-01-31 ec2059d
Update `pre-commit` to use `ruff` (fixes JosePizarro3 Q7)
