+++ pytaser/plotter.py
@@ -1,7 +1,5 @@
-import gettext
 import re
 import warnings
-from collections import defaultdict
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -25,11 +23,9 @@ def lambda_to_ev(lambda_float):
 
 def cutoff_transitions(dictionary, cutoff, ind_xmin, ind_xmax):
     """Output a list of transitions from a dict, with any that fall below a percentage cutoff of
-    the maximum value transition set to None."""
-    max_abs_val = {
-        key: np.max(abs(val[ind_xmin:ind_xmax]))
-        for key, val in dictionary.items()
-    }
+    the maximum value transition set to None.
+    """
+    max_abs_val = {key: np.max(abs(val[ind_xmin:ind_xmax])) for key, val in dictionary.items()}
     max_val = max(max_abs_val.values())
     relevant_transition_list = []
     for transition, value in max_abs_val.items():
@@ -41,8 +37,7 @@ def cutoff_transitions(dictionary, cutoff, ind_xmin, ind_xmax):
 
 
 class TASPlotter:
-    """
-    Class to generate a matplotlib plot of the TAS, DAS, or JDOS spectra, with a specific energy
+    """Class to generate a matplotlib plot of the TAS, DAS, or JDOS spectra, with a specific energy
     mesh, material, and conditions.
 
     Args:
@@ -87,8 +82,6 @@ class TASPlotter:
         self.weighted_jdos_dark_if = container.weighted_jdos_dark_if
         self.weighted_jdos_diff_if = container.weighted_jdos_diff_if
 
-        return None
-
     def fill_container_das(self, container):
         self.das_total = container.das_total
         self.alpha_new_sys = container.alpha_new_sys
@@ -112,7 +105,6 @@ class TASPlotter:
             self.bandgap_new_sys_lambda = ev_to_lambda(0.01)
         else:
             self.bandgap_new_sys_lambda = ev_to_lambda(self.bandgap_new_sys)
-        return None
 
     def get_plot(
         self,
@@ -126,12 +118,11 @@ class TASPlotter:
         yaxis="tas",
         **kwargs,
     ):
-        """
-        Plots TAS spectra using the data generated in the TAS container class as generated by
-        TASGenerator().generate_tas() or in the DAS container class generated by 
+        """Plots TAS spectra using the data generated in the TAS container class as generated by
+        TASGenerator().generate_tas() or in the DAS container class generated by
         DASGenerator().generate_das(). If the TASGenerator was not generated from VASP outputs,
         then the output TAS is generated using the change in joint density of states (JDOS) under
-        illumination, with no consideration of oscillator strengths ("Total TAS (ΔJDOS only)") –
+        illumination, with no consideration of oscillator strengths ("Total TAS (ΔJDOS only)") -
         this behaviour can also be explicitly chosen by setting "yaxis" to "jdos_diff".
 
         Otherwise, the output TAS is generated considering all contributions to the predicted TAS
@@ -139,19 +130,19 @@ class TASPlotter:
         desired, with stimulated emission neglected, set "yaxis" to "tas_absorption_only".
 
         One can also plot the JDOS before and after illumination, by setting "yaxis" to "jdos",
-        or the effective absorption coefficient (α, including absorption and stimulated emission)
+        or the effective absorption coefficient (a, including absorption and stimulated emission)
         before and after illumination, by setting "yaxis" to "alpha".
 
-        If a DAS container is parsed one can opt to set "yaxis" to "das" or "jdos_das". 
-        For the "yaxis=das" the difference between the optical absorption spectra of the new 
+        If a DAS container is parsed one can opt to set "yaxis" to "das" or "jdos_das".
+        For the "yaxis=das" the difference between the optical absorption spectra of the new
[…]
