+++ pytaser/generator.py
@@ -17,8 +17,7 @@ warnings.filterwarnings("ignore", category=RuntimeWarning)
 
 
 def gaussian(x, width, center=0.0, height=None):
-    """
-    Returns the values of gaussian(x) where x is array-like.
+    """Returns the values of gaussian(x) where x is array-like.
 
     Args:
         x: Input array.
@@ -34,8 +33,7 @@ def gaussian(x, width, center=0.0, height=None):
 
 
 def set_bandgap(bandstructure, dos, bandgap):
-    """
-    Shifts all bands of a material to correct the DFT-underestimated bandgap according to
+    """Shifts all bands of a material to correct the DFT-underestimated bandgap according to
     the input experimental bandgap.
 
     Args:
@@ -50,17 +48,13 @@ def set_bandgap(bandstructure, dos, bandgap):
     from copy import deepcopy
 
     if abs(dos.efermi - bandstructure.efermi) > 0.001:
-        raise ValueError(
-            "DOS and band structure are not from the same calculation"
-        )
+        raise ValueError("DOS and band structure are not from the same calculation")
 
     if bandstructure.is_metal():
         raise ValueError("System is a metal, cannot change bandgap")
 
     scissor = bandgap - bandstructure.get_band_gap()["energy"]
-    midgap = (
-        bandstructure.get_cbm()["energy"] + bandstructure.get_vbm()["energy"]
-    ) / 2
+    midgap = (bandstructure.get_cbm()["energy"] + bandstructure.get_vbm()["energy"]) / 2
 
     new_bandstructure = deepcopy(bandstructure)
     for spin, spin_energies in bandstructure.bands.items():
@@ -72,15 +66,13 @@ def set_bandgap(bandstructure, dos, bandgap):
     de = np.diff(dos.energies).mean()
     shift = int(scissor / 2 // de)
     new_dos = deepcopy(dos)
-    for spin in dos.densities.keys():
+    for spin in dos.densities:
         dens = np.zeros_like(dos.energies)
         if shift > 0:
             dens[: fermi_idx - shift] = dos.densities[spin][shift:fermi_idx]
             dens[fermi_idx + shift :] = dos.densities[spin][fermi_idx:-shift]
         else:
-            dens[abs(shift) : fermi_idx] = dos.densities[spin][
-                : fermi_idx + shift
-            ]
+            dens[abs(shift) : fermi_idx] = dos.densities[spin][: fermi_idx + shift]
             dens[fermi_idx:+shift] = dos.densities[spin][fermi_idx - shift :]
         new_dos.densities[spin] = dens
 
@@ -91,8 +83,7 @@ def set_bandgap(bandstructure, dos, bandgap):
 
 
 def jdos(bs, f, i, occs, energies, kweights, gaussian_width, spin=Spin.up):
-    """
-    Args:
+    """Args:
         bs: bandstructure object
         f: final band
         i: initial band
@@ -106,7 +97,6 @@ def jdos(bs, f, i, occs, energies, kweights, gaussian_width, spin=Spin.up):
         Cumulative JDOS value for a specific i->f transition, with consideration of
         partial occupancy and spin polarisation.
     """
-
     jdos = np.zeros(len(energies))
     for k in range(len(bs.bands[spin][i])):
         final_occ = occs[f][k]
@@ -116,26 +106,20 @@ def jdos(bs, f, i, occs, energies, kweights, gaussian_width, spin=Spin.up):
         k_weight = kweights[k]
         factor = k_weight * (init_occ - final_occ)
 
-        jdos += factor * gaussian(
-            energies, gaussian_width, center=final_energy - init_energy
-        )
+        jdos += factor * gaussian(energies, gaussian_width, center=final_energy - init_energy)
[…]
