+++ setup.py
@@ -1,5 +1,5 @@
 """"
-PyTASER: transient absorption prediction tool
+PyTASER: transient absorption prediction tool.
 """
 
 
