+++ pytaser/__init__.py
@@ -1,1 +1,1 @@
-"""PyTASER is a package for the generation, analysis and plotting of Transient Absorption Spectra of bulk materials."""
+"""PyTASER is a package for the generation, analysis and plotting of Transient Absorption Spectra (TAS)."""
