+++ pyproject.toml
@@ -54,3 +54,4 @@ ignore = [
 
 [tool.ruff.per-file-ignores]
 "tests/*" = ["D102", "D103"]
+"docs/*" = ["D100"]

[…]
