+++ pytaser/__init__.py
@@ -1,1 +1,2 @@
-"""PyTASER is a package for the generation, analysis and plotting of Transient Absorption Spectra (TAS)."""
+"""PyTASER is a package for the generation, analysis and plotting of Transient Absorption Spectra of bulk materials."""
+
