+++ pytaser/plotter.py
@@ -91,7 +91,8 @@ class TASPlotter:
         self.bandgap_ref = container.bandgap_ref
         if self.bandgap_ref <= 0.01:
             print(
-                "Band gap < 0.01. Settiing to 0.01 eV to avoid division by zero when converting energies from eV to wavelength in nm."
+                "Band gap < 0.01. Settiing to 0.01 eV to avoid division by zero when converting energies from eV to "
+                "wavelength in nm. "
             )
             self.bandgap_ref_lambda = ev_to_lambda(0.01)
         else:
@@ -100,7 +101,8 @@ class TASPlotter:
         self.bandgap_new_sys = container.bandgap_new_sys
         if self.bandgap_new_sys <= 0.01:
             print(
-                "Band gap < 0.01. Settiing to 0.01 eV to avoid division by zero when converting energies from eV to wavelength in nm."
+                "Band gap < 0.01. Settiing to 0.01 eV to avoid division by zero when converting energies from eV to "
+                "wavelength in nm. "
             )
             self.bandgap_new_sys_lambda = ev_to_lambda(0.01)
         else:
@@ -116,6 +118,7 @@ class TASPlotter:
         ymin=None,
         ymax=None,
         yaxis="tas",
+        invert_axis=False,
         **kwargs,
     ):
         """Plots TAS spectra using the data generated in the TAS container class as generated by
@@ -170,20 +173,18 @@ class TASPlotter:
                 joint density of states (JDOS) before & after illumination ("Total TAS (ΔJDOS
                 only)") - the latter can also be explicitly selected with yaxis = "jdos_diff".
 
-                If yaxis = "tas_absorption_only", will instead plot the predicted TAS spectrum
-                considering only the change in absorption, with stimulated emission neglected.
-                If yaxis = "jdos", will plot the joint density of states before & after
-                illumination.
-                If yaxis = "alpha", will plot the effective absorption coefficient (a, including
-                absorption and stimulated emission) before and after illumination.
-                If yaxis = "das", will plot the difference between the optical absorption spectra
-                of the new system and the reference system, together with the absorption spectra
-                If yaxis = "jdos_das", will plot the difference between the JDOS of the new system
-                and the reference system, together with the JDOS of both systems.
-            **kwargs: Additional arguments to be passed to matplotlib.pyplot.legend(); such as
-                `ncols` (number of columns in the legend), `loc` (location of the legend),
-                `fontsize` etc. (see https://matplotlib.org/stable/api/_as_gen/matplotlib.pyplot
-                .legend.html)
+                If yaxis = "tas_absorption_only", will instead plot the predicted TAS spectrum considering only the
+                change in absorption, with stimulated emission neglected. If yaxis = "jdos", will plot the joint
+                density of states before & after illumination. If yaxis = "alpha", will plot the effective absorption
+                coefficient (a, including absorption and stimulated emission) before and after illumination. If yaxis
+                = "das", will plot the difference between the optical absorption spectra of the new system and the
+                reference system, together with the absorption spectra If yaxis = "jdos_das", will plot the
+                difference between the JDOS of the new system and the reference system, together with the JDOS of
+                both systems. invert_axis: If true, x-axis decreases from left-right. If false (default), x-axis will
+                increase from left-right. Especially relevant when switching between wavelength and electronvolts.
+                **kwargs: Additional arguments to be passed to matplotlib.pyplot.legend(); such as `ncols` (number of
+                columns in the legend), `loc` (location of the legend), `fontsize` etc. (see
+                https://matplotlib.org/stable/api/_as_gen/matplotlib.pyplot .legend.html)
 
[…]
