+++ requirements.txt
@@ -1,8 +1,8 @@
 pytest>=7.1.3
-pytest-mpl==0.15.1
+pytest-mpl==0.39.5
 numpy
 monty
-pymatgen>=2023.05.31
+pymatgen>=2024.1.27
 matplotlib>=3.7.1
 scipy
 pathlib
