+++ paper/paper.bib
@@ -228,7 +228,7 @@ Program Title: Abinit Program Files doi: http://dx.doi.org/10.17632/csvdrr4d68.1
 @article{huang2021perovskite,
   title={Perovskite-inspired materials for photovoltaics and beyond—from design to devices},
   author={Huang, Yi-Teng and Kavanagh, Se{\'a}n R and Scanlon, David O and Walsh, Aron and Hoye, Robert LZ},
-  doi={10.1088/1361-6528/abcf6d}
+  doi={10.1088/1361-6528/abcf6d},
   journal={Nanotechnology},
   volume={32},
   number={13},
