+++ README.md
@@ -32,12 +32,12 @@ PyTASER is designed for users with moderate experience in computational methods 
 
 * Use of Python as the programming language (due to its low entry barrier, flexibility and popularity in the materials modelling field)
 * Documentation and easy-to-follow workflows with complete unit-test coverage.
-* Interfaced with the popular materials analysis package `pymatgen <https://pymatgen.org/index.html>`__
-* Currently compatible with VASP, the most popular electronic structure calculation code. Other codes will be added soon.
+* Interfaced with the popular materials analysis package [`pymatgen`](https://pymatgen.org/index.html).
+* Currently compatible with VASP, while support for other electronic structure codes is planned – if you have a specific request, please open an issue!
 
 # Installation
 
-To install the module with pip (recommended): 
+To install the module with `pip` (recommended): 
 
 ```
 pip install pytaser
@@ -74,11 +74,8 @@ This command tries to obtain the required packages and their dependencies and in
 
 # Visualisation 
 
-One approach is to use a Jupyter Notebook, as shown in the [examples](https://github.com/WMD-group/PyTASER/blob/main/examples) folder.
-Alternatively, you can set up a file in Python to run in the command line of the terminal:
-```
-python3 <filename.py>
-```
+The recommended approach is to use ``PyTASER`` within an interactive python environment (such as a Jupyter Notebook), as shown in the [tutorials](https://pytaser.readthedocs.io/en/latest/dft_examples.html) on the docs and the [examples](https://github.com/WMD-group/PyTASER/blob/main/examples) folder.
+Alternatively, you can use ``PyTASER`` through python scripts (e.g. with `python <filename.py>`).
 
 If using your own, locally calculated data, please follow [this](https://github.com/WMD-group/PyTASER/blob/main/examples/PyTASER_DFT_Example.ipynb) workflow.
 Otherwise, use the MP-integrated workflow [here](https://github.com/WMD-group/PyTASER/blob/main/examples/PyTASER_MP_Example.ipynb). 
