### 2024-02-07 73666a1
Remove unnecessary print statements, update `get_plot()` docstrings, formatting
