+++ pytaser/plotter.py
@@ -1,3 +1,7 @@
+"""
+TAS Plotting Tools.
+"""
+
 import re
 import warnings
 
@@ -39,12 +43,6 @@ def cutoff_transitions(dictionary, cutoff, ind_xmin, ind_xmax):
 class TASPlotter:
     """Class to generate a matplotlib plot of the TAS, DAS, or JDOS spectra, with a specific energy
     mesh, material, and conditions.
-
-    Args:
-        container: TAS container class as generated by TASGenerator().generate_tas() or DASGenerator().generate_das()
-        material_name: Name of material being investigated (string) [optional, for labelling]
-        system_name: Name of the new system used for DAS (string) [optional, for labelling]
-        reference_name: Name of the reference system used for DAS (string) [optional, for labelling]
     """
 
     def __init__(
@@ -54,6 +52,21 @@ class TASPlotter:
         system_name=None,
         reference_name=None,
     ):
+        """
+        Args:
+            container:
+                TAS container class as generated by TASGenerator().generate_tas() or
+                DASGenerator().generate_das()
+            material_name: Name of material being investigated (string) [optional, for labelling]
+            system_name: Name of the new system used for DAS (string) [optional, for labelling]
+            reference_name: Name of the reference system used for DAS (string) [optional, for labelling].
+
+        Args:
+            container:
+            material_name:
+            system_name:
+            reference_name:
+        """
         self.energy_mesh_ev = container.energy_mesh_ev
         self.material_name = material_name
         self.reference_name = reference_name
@@ -62,11 +75,11 @@ class TASPlotter:
         self.temp = container.temp
 
         if isinstance(container, Tas):
-            self.fill_container_tas(container)
+            self._fill_container_tas(container)
         else:
-            self.fill_container_das(container)
+            self._fill_container_das(container)
 
-    def fill_container_tas(self, container):
+    def _fill_container_tas(self, container):
         self.tas_total = container.tas_total
         self.jdos_diff_if = container.jdos_diff_if
         self.jdos_light_total = container.jdos_light_total
@@ -82,7 +95,7 @@ class TASPlotter:
         self.weighted_jdos_dark_if = container.weighted_jdos_dark_if
         self.weighted_jdos_diff_if = container.weighted_jdos_diff_if
 
-    def fill_container_das(self, container):
+    def _fill_container_das(self, container):
         self.das_total = container.das_total
         self.alpha_new_sys = container.alpha_new_sys
         self.alpha_ref = container.alpha_ref
@@ -90,20 +103,14 @@ class TASPlotter:
         self.jdos_ref_total = container.jdos_ref_total
         self.bandgap_ref = container.bandgap_ref
         if self.bandgap_ref <= 0.01:
-            print(
-                "Band gap < 0.01. Settiing to 0.01 eV to avoid division by zero when converting energies from eV to "
-                "wavelength in nm. "
-            )
+            # bandgap <0.01 eV (likely metallic), so set to 0.01 eV to avoid division by zero
             self.bandgap_ref_lambda = ev_to_lambda(0.01)
         else:
             self.bandgap_ref_lambda = ev_to_lambda(self.bandgap_ref)
 
         self.bandgap_new_sys = container.bandgap_new_sys
         if self.bandgap_new_sys <= 0.01:
-            print(
-                "Band gap < 0.01. Settiing to 0.01 eV to avoid division by zero when converting energies from eV to "
-                "wavelength in nm. "
-            )
+            # bandgap <0.01 eV (likely metallic), so set to 0.01 eV to avoid division by zero
             self.bandgap_new_sys_lambda = ev_to_lambda(0.01)
         else:
             self.bandgap_new_sys_lambda = ev_to_lambda(self.bandgap_new_sys)
[…]
