+++ pytaser/generator.py
@@ -1,3 +1,7 @@
+"""
+This module contains the TASGenerator class, which is used to generate TAS spectra.
+"""
+
 import warnings
 from multiprocessing import Array, Pool, cpu_count
 
@@ -17,7 +21,9 @@ warnings.filterwarnings("ignore", category=RuntimeWarning)
 
 
 def gaussian(x, width, center=0.0, height=None):
-    """Returns the values of gaussian(x) where x is array-like.
+    """
+    Returns Gaussian curve(s) centred at point(s)
+    x, where x is array-like.
 
     Args:
         x: Input array.
@@ -219,8 +225,8 @@ def get_nonzero_band_transitions(
     )
 
 
-def init_shared_memory(cder, occs, eigs_shifted, norm_kweights):
-    global _cder, _occs, _eigs_shifted, _norm_kweights
+def _init_shared_memory(cder, occs, eigs_shifted, norm_kweights):
+    global _cder, _occs, _eigs_shifted, _norm_kweights  # pylint: disable=global-statement
     global _cder_shape, _occs_shape, _eigs_shifted_shape, _norm_kweights_shape
 
     _cder_shape = cder.shape
@@ -327,7 +333,7 @@ def occ_dependent_alpha(
     ):  # quicker without multiprocessing below this arg length
         with Pool(
             processes,
-            initializer=init_shared_memory,
+            initializer=_init_shared_memory,
             initargs=shared_memory_args,
         ) as pool:
             results = pool.map(
@@ -391,26 +397,28 @@ def get_cbm_vbm_index(bs):
 
 
 class TASGenerator:
-    """Class to generate a TAS spectrum (decomposed and cumulative) from a bandstructure and
-    dos object.
-
-    Args:
-        bs: Pymatgen-based bandstructure object
-        kpoint_weights: kpoint weights either found by the function or inputted.
-        dos: Pymatgen-based dos object
-        dfc: Pymatgen-based DielectricFunctionCalculator object (for computing oscillator strengths)
-
-    Attributes:
-        bs: Pymatgen bandstructure object
-        kpoint_weights: k-point weights (degeneracies).
-        dos: Pymatgen-based dos object
-        dfc: Pymatgen-based DielectricFunctionCalculator object (for computing oscillator strengths)
-        bg_centre: Energy (eV) of the bandgap centre.
-        vb: Spin dict detailing the valence band maxima.
-        cb: Spin dict detailing the conduction band minima
+    """
+    Class to generate a TAS spectrum (decomposed and cumulative) from
+    a bandstructure and dos object.
     """
 
     def __init__(self, bs, kpoint_weights, dos, dfc=None):
+        """
+        Args:
+            bs: Pymatgen-based bandstructure object
+            kpoint_weights: kpoint weights either found by the function or inputted.
+            dos: Pymatgen-based dos object
+            dfc: Pymatgen-based DielectricFunctionCalculator object (for computing oscillator strengths).
+
+        Attributes:
+            bs: Pymatgen bandstructure object
+            kpoint_weights: k-point weights (degeneracies).
+            dos: Pymatgen-based dos object
+            dfc: Pymatgen-based DielectricFunctionCalculator object (for computing oscillator strengths)
+            bg_centre: Energy (eV) of the bandgap centre.
+            vb: Spin dict detailing the valence band maxima.
+            cb: Spin dict detailing the conduction band minima
+        """
         self.bs = bs
         self.kpoint_weights = kpoint_weights
         self.dos = FermiDos(dos)
@@ -439,16 +447,17 @@ class TASGenerator:
         """
         warnings.filterwarnings("ignore", category=UnknownPotcarWarning)
         warnings.filterwarnings("ignore", message="No POTCAR file with matching TITEL fields")
-        vr = Vasprun(vasprun_file)
[…]
