### 2024-02-07 cd755da
Update tutorials; switch to  usage, update labels explanation, update plots, general cleanup
