### 2024-02-07 a18ec22
docs update with theory
