+++ README.md
@@ -22,11 +22,6 @@ The goal of this library is to simulate spectra for comparison with and interpre
 * Integration with the Materials Project database, allowing support for non-locally calculated materials
 * Ability to produce publication-ready figures, with flexibility in plotting.
 
-# Background
-
-TAS is a powerful pump-probe tool to characterise the excited states of materials, while DAS can be used to represent how changes in a system affects its ground-state optical absorption spectra. These techniques can be used to understand microscopic processes in photochemical and electrochemical transformations, including phenomena such as electron trapping and carrier recombination.
-
-The drawback is that TAS spectra are difficult to interpret, especially for crystals where the specific valence and conduction band structure can give rise to complex features. Our goal here is to predict TAS features from first-principles starting from the most simple models of static excitations through to the kinetics of relaxation of the excited state back to the ground state.
 
 PyTASER is designed for users with moderate experience in computational methods and optical techniques, enabled by the following features:
 
@@ -34,6 +29,24 @@ PyTASER is designed for users with moderate experience in computational methods 
 * Documentation and easy-to-follow workflows with complete unit-test coverage.
 * Interfaced with the popular materials analysis package [`pymatgen`](https://pymatgen.org/index.html).
 * Currently compatible with VASP, while support for other electronic structure codes is planned – if you have a specific request, please open an issue!
+
+# Background
+
+TAS is a powerful pump-probe tool to characterise the excited states of materials, while DAS can be used to represent how changes in a system affects its ground-state optical absorption spectra. These techniques can be used to understand microscopic processes in photochemical and electrochemical transformations, including phenomena such as electron trapping and carrier recombination.
+
+The drawback is that TAS spectra are difficult to interpret, especially for crystals where the specific valence and conduction band structure can give rise to complex features. Our goal here is to predict TAS features from first-principles starting from the most simple models of static excitations through to the kinetics of relaxation of the excited state back to the ground state.
+
+To achieve this, PyTASER identifies the allowed vertical optical transitions between electronic bands of the material to determine possible excitations that can occur from the ground 'dark' and excited 'light' electronic states.
+This is done by calculating the effective absorption in each state - this is a product of the joint density of states (JDOS) and the transition probability for each band transition, both of which are based on post-processing ground state DFT calculations. Once calculated, PyTASER compares changes in electronic transitions between the dark and light states, as demonstrated in the figure below.
+
+<p align="center">
+  <img src="docs/source/_static/jdos-tas-schematic.png" alt="Schematic TAS">
+</p>
+<p align="center">
+  <em>Schematics of the ground and excited state electronic structures and optical profiles. The ground 'dark' state is at the top, showing full occupancy and unoccupancy (blue, orange) for the conduction and valence bands respectively. The excited 'light' state shows partial occupancy in a similar plot at the bottom. The overall DA plot is displayed to the right, the difference between the dark and light effective absorption plots.</em>
+</p>
+
+## JDOS method
 
 # Installation
 
