### 2024-02-07 b47f522
Revert "update setup.py"
This reverts commit 7ae40fd7b6bd87bd633d2c778e99d8f6b902d9f2.
