### 2024-02-08 715a6cd
Update docs requirements and `.gitignore`
