### 2024-02-08 dbe30a0
Update example MP notebook with warning about non-uniform meshes
