### 2024-02-08 810c20c
Swithch docs tutorial generation to `myst_nb`, shows note/warning boxes etc, and update version number in `conf.py`
