### 2024-02-08 653a4ea
Add `myst-nb` requirement
