### 2024-02-08 #5999 @savya10
> > Hi @savya10 The author list on the zenodo archive and the one is the paper kinda match, but there are a lot of variations like "Youngwon" on the zenodo site and "Young Won" on the paper. Can update one or the other of these? Probably easier to update the paper?
>
> Sure - will update the zenodo as I just remembered that Young Won is the correct way

done!
