### 2024-02-12 #5999 @Kevin-Mattheus-Moerman
@savya10 as AEiC I will now assist in the final steps to process paper acceptance. I have checked the paper, the archive link, this review, and your repository. Most seems in order, I only have the below minor points which requires your attention (no new version/archive is needed):

Typos:
- [ ] Check spelling for `receieved` which should read `received`
