+++ pytaser/das_generator.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 """Created on Thu Aug  3 16:42:36 2023.
 
 @author: lucasverga
@@ -32,6 +31,12 @@ class DASGenerator:
         new_system,
         reference_system,
     ):
+        """
+
+        Args:
+            new_system: Internal_Abs object from internal_abs_generator for the new system
+            reference_system: Internal_Abs object from internal_abs_generator for the reference system.
+        """
         self.new_system = new_system
         self.reference_system = reference_system
 
@@ -55,7 +60,8 @@ class DASGenerator:
             waveder_file_ref: The WAVEDER file for the reference system.
 
         Returns:
-            A DASGenerator object containing the Internal_Abs object for the new system and reference system.
+            A DASGenerator object containing the Internal_Abs object for the new system and reference
+            system.
         """
         warnings.filterwarnings("ignore", category=UnknownPotcarWarning)
         warnings.filterwarnings("ignore", message="No POTCAR file with matching TITEL fields")
@@ -91,7 +97,8 @@ class DASGenerator:
             mpr_ref: An MPRester object for the reference system if already generated by user.
 
         Returns:
-            A DASGenerator object containing the Internal_Abs object for the new system and reference system.
+            A DASGenerator object containing the Internal_Abs object for the new system and reference
+            system.
         """
         new_system = Internal_Abs.internal_from_mpid(mpid, bg=None, api_key=None, mpr=None)
         reference_system = Internal_Abs.internal_from_mpid(mpid_ref, bg_ref, api_key=None, mpr_ref=None)
