+++ setup.py
@@ -1,6 +1,4 @@
-""""
-PyTASER: transient absorption prediction tool.
-"""
+"""PyTASER: transient absorption prediction tool."""
 
 
 import pathlib
