### 2024-02-14 af4e2b4
Update linting and formatting
