+++ pytaser/plotter.py
@@ -16,18 +16,23 @@ warnings.filterwarnings("ignore", category=RuntimeWarning)
 
 
 def ev_to_lambda(ev):
-    """Convert photon energies from eV to a wavelength in nm."""
+    """
+    Convert photon energies from eV to a wavelength in nm.
+    """
     return ((scpc.h * scpc.c) / (ev * scpc.electron_volt)) * 10e8
 
 
 def lambda_to_ev(lambda_float):
-    """Convert photon energies from a wavelength in nm to eV."""
+    """
+    Convert photon energies from a wavelength in nm to eV.
+    """
     return (10e8 * (scpc.h * scpc.c)) / (lambda_float * scpc.electron_volt)
 
 
 def cutoff_transitions(dictionary, cutoff, ind_xmin, ind_xmax):
-    """Output a list of transitions from a dict, with any that fall below a percentage cutoff of
-    the maximum value transition set to None.
+    """
+    Output a list of transitions from a dict, with any that fall below a
+    percentage cutoff of the maximum value transition set to None.
     """
     max_abs_val = {key: np.max(abs(val[ind_xmin:ind_xmax])) for key, val in dictionary.items()}
     max_val = max(max_abs_val.values())
@@ -41,8 +46,9 @@ def cutoff_transitions(dictionary, cutoff, ind_xmin, ind_xmax):
 
 
 class TASPlotter:
-    """Class to generate a matplotlib plot of the TAS, DAS, or JDOS spectra, with a specific energy
-    mesh, material, and conditions.
+    """
+    Class to generate a matplotlib plot of the TAS, DAS, or JDOS spectra, with
+    a specific energy mesh, material, and conditions.
     """
 
     def __init__(
@@ -122,12 +128,16 @@ class TASPlotter:
         invert_axis=False,
         **kwargs,
     ):
-        """Plots TAS spectra using the data generated in the TAS container class as generated by
-        TASGenerator().generate_tas() or in the DAS container class generated by
-        DASGenerator().generate_das(). If the TASGenerator was not generated from VASP outputs,
-        then the output TAS is generated using the change in joint density of states (JDOS) under
-        illumination, with no consideration of oscillator strengths ("Total TAS (ΔJDOS only)") -
-        this behaviour can also be explicitly chosen by setting "yaxis" to "jdos_diff".
+        """
+        Plots TAS spectra using the data generated in the TAS container class
+        as generated by TASGenerator().generate_tas() or in the DAS container
+        class generated by DASGenerator().generate_das(). If the TASGenerator
+        was not generated from VASP outputs, then the output TAS is generated
+        using the change in joint density of states (JDOS) under.
+
+        illumination, with no consideration of oscillator strengths ("Total TAS
+        (ΔJDOS only)") - this behaviour can also be explicitly chosen by
+        setting "yaxis" to "jdos_diff".
 
         Otherwise, the output TAS is generated considering all contributions to the predicted TAS
         spectrum ("Total TAS (Δɑ)"). If only the contribution from the change in absorption is
