+++ pytaser/__init__.py
@@ -1,3 +1,4 @@
-"""PyTASER is a package for the generation, analysis and plotting of Transient Absorption Spectra of bulk
-materials.
+"""
+PyTASER is a package for the generation, analysis and plotting of Transient
+Absorption Spectra of bulk materials.
 """
