### 2024-02-14 cec285c
Add citation file and links to JOSS paper for deeper theory as discussed
