### 2024-02-14 5b8a6d5
Update `releasing.yml` to only trigger when `setup.py` updated (i.e. new version number)
